// database/prisma-oppulence/generated/zod/inputTypeSchemas/DealProductCountOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const DealProductCountOrderByAggregateInputSchema: z.ZodType<Prisma.DealProductCountOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        dealId: z.lazy(() => SortOrderSchema).optional(),
        productId: z.lazy(() => SortOrderSchema).optional(),
        quantity: z.lazy(() => SortOrderSchema).optional(),
        unitPrice: z.lazy(() => SortOrderSchema).optional(),
        discount: z.lazy(() => SortOrderSchema).optional(),
        total: z.lazy(() => SortOrderSchema).optional(),
        description: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        updatedAt: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default DealProductCountOrderByAggregateInputSchema